prefect/requirements.txt
prefect==2.11.3
prefect-shell==0.1.5
boto3==1.28.23
python-dotenv==1.0.0
s3fs==0.4.2
pandas==2.0.3
dbt-core==1.6.2
dbt-bigquery==1.6.6
google-auth==2.14.1
google-api-core==2.12.0
google-api-python-client==2.65.0